requirements.txt
pyspark
pandas
pyarrow
lightgbm
scikit-learn
typer
pydantic>=2
SQLAlchemy
duckdb
python-dotenv
apache-airflow